requirements.txt
streamlit==1.52.2
psycopg[binary]
Pillow
streamlit_js_eval
